Shoot'n egg'n and freez'n blocks, (...)
Add WorldInteractions (Registry) that reads data/arcana/interactions/freeze.json and uses it to get blocks that can be converted if used IceCast.

To add/remove blocks that can be converted, add to values JsonObject that contains "from", "to" and optionally "cover" (what block is placed on top of converted block).

Documentation in commit enjoy :)